# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tasing API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Tasing API Server"
        }
    ],
    "paths": {
...
        "/api/dashboard/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Display the authenticated user's Profile",
                "operationId": "show",
                "responses": {
                    "200": {
                        "description": "Profile retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update the authenticated user's Profile with given data",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastname",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone_number1",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "address_line1",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "address_line2",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "old_password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "422": {
                        "description": "The given data was invalid."
                    }
                },
                "security": [
                    {
                        "": []
                    }
                ]
            }
        },
        "/api/dashboard/multi-api/index": {
            "get": {
                "tags": [
                    "Scrapings APIs"
                ],
                "summary": "Display the list of projects (scraping projects)",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Projects data retrived.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "filters": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "field-1": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "": []
                    }
                ]
            }
        },
        "/api/dashboard/multi-api/get_front_info": {
            "get": {
                "tags": [
                    "Scrapings APIs"
                ],
                "summary": "Get the front info for given project.",
                "operationId": "getFrontInfo",
                "parameters": [
                    {
                        "name": "project",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Projects retrived successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Project not found."
                    },
                    "422": {
                        "description": "The given data was invalid."
                    }
                },
                "security": [
                    {
                        "": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "description": "id",
                        "type": "integer",
                        "format": "int32"
                    },
                    "name": {
                        "description": "name",
                        "type": "string"
                    },
                    "lastname": {
                        "description": "lastname",
                        "type": "string"
                    },
                    "email": {
                        "description": "email",
                        "type": "string"
                    },
                    "phone_number1": {
                        "description": "phone_number1",
                        "type": "string"
                    },
                    "phone_number2": {
                        "description": "phone_number2",
                        "type": "string"
                    },
                    "country_code": {
                        "description": "country_code",
                        "type": "string"
                    },
                    "address_line1": {
                        "description": "address_line1",
                        "type": "string"
                    },
                    "address_line2": {
                        "description": "address_line2",
                        "type": "string"
                    },
                    "city": {
                        "description": "city",
                        "type": "string"
                    },
                    "region": {
                        "description": "region",
                        "type": "string"
                    },
                    "zipcode": {
                        "description": "zipcode",
                        "type": "string"
                    },
                    "company_name": {
                        "description": "company_name",
                        "type": "string"
                    },
                    "company_number": {
                        "description": "company_number",
                        "type": "string"
                    },
                    "email_verified_at": {
                        "description": "email_verified_at",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    }
}